requirements.txt
# Telegram Bot
python-telegram-bot>=21.0

# Async SQLite
aiosqlite>=0.19.0

# HTTP client for APIs
aiohttp>=3.9.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
aioresponses>=0.7.0

# Linters and Code Quality
flake8>=6.0.0
black>=23.0.0
mypy>=1.0.0
isort>=5.12.0
